Moved on-disk disassembly generation to package
