+++ external-crates/move/crates/move-cli/src/base/test.rs
@@ -6,7 +6,6 @@ use crate::NativeFunctionRecord;
 use anyhow::Result;
 use clap::*;
 use move_binary_format::CompiledModule;
-use move_bytecode_source_map::utils::serialize_to_json_string;
 use move_command_line_common::files::MOVE_COVERAGE_MAP_EXTENSION;
 use move_compiler::{
     diagnostics::{self, Diagnostics},
@@ -15,24 +14,10 @@ use move_compiler::{
     PASS_CFGIR,
 };
 use move_coverage::coverage_map::{output_map_to_file, CoverageMap};
-use move_disassembler::disassembler::Disassembler;
-use move_package::{
-    compilation::{
-        build_plan::BuildPlan,
-        compiled_package::{CompiledUnitWithSource, OnDiskCompiledPackage, OnDiskPackage},
-        package_layout::CompiledPackageLayout,
-    },
-    BuildConfig,
-};
-use move_symbol_pool::Symbol;
+use move_package::{compilation::build_plan::BuildPlan, BuildConfig};
 use move_unit_test::UnitTestingConfig;
 use move_vm_test_utils::gas_schedule::CostTable;
-use std::{
-    collections::BTreeSet,
-    io::Write,
-    path::{Path, PathBuf},
-    process::ExitStatus,
-};
+use std::{io::Write, path::Path, process::ExitStatus};
 // if windows
 #[cfg(target_family = "windows")]
 use std::os::windows::process::ExitStatusExt;
@@ -167,6 +152,7 @@ pub fn run_move_unit_tests<W: Write + Send>(
     let mut test_plan = None;
     build_config.test_mode = true;
     build_config.dev_mode = true;
+    build_config.save_disassembly = save_disassembly;
 
     // Build the resolution graph (resolution graph diagnostics are only needed for CLI commands so
     // ignore them by passing a vector as the writer)
@@ -211,7 +197,7 @@ pub fn run_move_unit_tests<W: Write + Send>(
     // then save it, before resuming the rest of the compilation and returning the results and
     // control back to the Move package system.
     let mut warning_diags = None;
-    let compiled_package = build_plan.compile_with_driver(writer, |compiler| {
+    build_plan.compile_with_driver(writer, |compiler| {
         let (files, comments_and_compiler_res) = compiler.run::<PASS_CFGIR>().unwrap();
         let (_, compiler) =
             diagnostics::unwrap_or_report_pass_diagnostics(&files, comments_and_compiler_res);
@@ -272,56 +258,7 @@ pub fn run_move_unit_tests<W: Write + Send>(
         let coverage_map = CoverageMap::from_trace_file(trace_path);
         output_map_to_file(coverage_map_path, &coverage_map).unwrap();
     }
-    if save_disassembly {
-        let build_dir_path = pkg_path.join(CompiledPackageLayout::Root.path());
-        let on_disk_package = OnDiskCompiledPackage {
-            root_path: build_dir_path.join(root_package.as_str()),
-            package: OnDiskPackage {
-                compiled_package_info: compiled_package.compiled_package_info.clone(),
-                dependencies: compiled_package
-                    .deps_compiled_units
-                    .iter()
-                    .map(|(package_name, _)| *package_name)
-                    .collect::<BTreeSet<_>>()
-                    .into_iter()
-                    .collect(),
-            },
-        };
-        for compiled_unit in &compiled_package.root_compiled_units {
-            save_disassembly_to_disk(&on_disk_package, root_package, compiled_unit)?;
-        }
-        for (dep_name, compiled_unit) in &compiled_package.deps_compiled_units {
-            save_disassembly_to_disk(&on_disk_package, *dep_name, compiled_unit)?;
-        }
-    }
     Ok((UnitTestResult::Success, warning_diags))
-}
-
-fn save_disassembly_to_disk(
-    compiled_package: &OnDiskCompiledPackage,
-    package_name: Symbol,
-    unit: &CompiledUnitWithSource,
-) -> Result<()> {
-    let root_package = compiled_package.package.compiled_package_info.package_name;
-    let bytecode_modules_dir = CompiledPackageLayout::CompiledModules.path();
-    let file_path = if root_package == package_name {
-        PathBuf::new()
-    } else {
-        CompiledPackageLayout::Dependencies
-            .path()
-            .join(package_name.as_str())
-    }
-    .join(unit.unit.name.as_str());
-    let d = Disassembler::from_unit(&unit.unit);
-    let (disassembled_string, bytecode_map) = d.disassemble_with_source_map()?;
-    compiled_package.save_under(
-        bytecode_modules_dir.join(&file_path).with_extension("mvb"),
-        disassembled_string.as_bytes(),
-    )?;
-    compiled_package.save_under(
-        bytecode_modules_dir.join(&file_path).with_extension("json"),
-        serialize_to_json_string(&bytecode_map)?.as_bytes(),
-    )
 }
 
 impl From<UnitTestResult> for ExitStatus {
