+++ external-crates/move/crates/move-package/src/lib.rs
@@ -60,6 +60,11 @@ pub struct BuildConfig {
     #[clap(name = "generate-docs", long = "doc", global = true)]
     pub generate_docs: bool,
 
+    /// Save disassembly for generated bytecode along with
+    /// bytecode maps (source maps for disassembeld bytecode)
+    #[clap(name = "save-disassembly", long = "disassemble", global = true)]
+    pub save_disassembly: bool,
+
     /// Installation directory for compiled artifacts. Defaults to current directory.
     #[clap(long = "install-dir", global = true)]
     pub install_dir: Option<PathBuf>,
